requirements.txt
Pillow==10.0.0
Pillow==10.3.0
streamlit==1.30.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
